requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
bitsandbytes==0.46.0
bleach==6.2.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
grpcio==1.73.1
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kaggle==1.7.4.5
kiwisolver==1.4.7
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.9.4
modelscope==1.27.1
mpmath==1.3.0
multidict==6.6.2
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
nvidia-ml-py==12.575.51
packaging==25.0
pandas==2.3.0
peft==0.15.2
pillow==11.2.1
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyecharts==2.0.8
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.4
rouge==1.0.1
rouge_score==0.1.2
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
swankit==0.2.4
swanlab==0.6.4
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.8.0.dev20250624+cu128
torchaudio==2.8.0.dev20250625+cu128
torchvision==0.23.0.dev20250625+cu128
tqdm==4.67.1
transformers==4.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unsloth==2024.8
urllib3==1.26.20
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
